Assign sensitive detector property to volume before the actual one.
In the original nexus implementation, the sensitive detector we use to
store sensor hits (PmtSD) needed to be assigned to
the volume *before* the active one, in order for the process that counts
hits to be triggered. In commits 00b093a and 74222f7, we changed this
behaviour to the more logic one of assigning the sensitive detector
property to the volume actually detecting photons, because both
implementations produced the same result. However, with
Geant4.10.7, it seems that this implementation doesn't work anymore,
so we have changed the code back to the original one. Tests that check
that the behaviour is correct for DEMO++, NEW, NEXT-100 and Flex100 geometry succeed.