Remove the possibility of not placing the lead castle
The reason is that this possibility creates an extra layer of
complexity for vertex generation, with more possibilities to make mistakes.
A consequence of this is the removal of the axial external generators
and collimator because they didn't fit with the castle (and were the
reason why we added the possibility of not placing the castle at
all). However, we are not going to simulate external calibrations anymore,
since NEW has been dismantled, therefore it should not be a problem now.